# openapi/authentication/authentication.yaml
openapi: 3.0.0
info:
  title: Hestix API
  description: Security endpoints for the Hestix API
  version: 1.0.0
paths:
  /auth/register:
    post:
      tags:
        - Auth
      operationId: registerUser
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationResponse'

  /auth/token:
    post:
      tags:
        - Auth
      operationId: loginUser
      summary: Authenticate user and get token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: Token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

components:
  schemas:
    AuthenticationRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [ username, password ]

    TokenResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - refreshExpiresIn
        - tokenType
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
        refreshExpiresIn:
          type: integer
          format: int64
        tokenType:
          type: string

    UserRegistrationRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required: [ username, email, password ]

    UserRegistrationResponse:
      type: object
      required:
        - username
        - email
        - token
      properties:
        username:
          type: string
        email:
          type: string
        token:
          $ref: '#/components/schemas/TokenResponse'